requirements.txt
# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.0
pydantic-settings==2.1.0

# 配置和数据处理
pyyaml==6.0.1
python-multipart==0.0.6

# 模板引擎
jinja2==3.1.2

# 安全相关
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP客户端（用于代理功能）
requests==2.31.0

# 工具库
python-dotenv==1.0.0

# 开发依赖
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0